classify/classify/requirements.txt
django
opencv-python
scikit-image
keras
Image
numpy
path.py
tensorflow